+++ prototype/core/hss/parsing/HSSNumberConstant.h
@@ -43,10 +43,10 @@
  *
  *      FILE INFORMATION:
  *      =================
- *      Last changed: 2011/04/25
+ *      Last changed: 2011/06/12
  *      HSS version: 1.0
  *      Core version: 0.3
- *      Revision: 5
+ *      Revision: 6
  *
  ********************************************************************/
 
@@ -66,12 +66,12 @@ namespace AXR {
         virtual ~HSSNumberConstant();
         
         void setValue(long double newValue);
-        double getValue();
+        long double getValue();
         
         std::string toString();
         
     protected:
-        double value;
+        long double value;
     };
 }
